Got rid of Duration
The duration column was causing too many issues with the models, so it has been removed. I have left the old cleaning and model notebooks (the ones in which duration was used) as an example and lesson for myself.

The new cleaning notebook goes through and manually changes everything to categories, the new models notebook utilises SVM, Decision Tree and random forest. I will also do corss vlaidation of both SVM and decision tree